+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## 1.7.6
+25/09/2021
+* Added support for "Master" type.
+
 ## 1.7.5
 18/08/2021
-* bug fix - records handling failing after update 1.5
+* Bug fix - records handling failing after update 1.5
 * Added Facebook XPATH as a config option
 
 ## 1.7.4